Cms 1

See merge request hparichha/fashionhubbackend!20

+++ catalog-service/src/main/java/com/catalogservice/dto/ProductCategoryRequestDto.java
@@ -2,19 +2,22 @@ package com.catalogservice.dto;
 
 import lombok.Data;
 
+import java.math.BigInteger;
+
 @Data
 public class ProductCategoryRequestDto {
     private String sku;          // SKU of the product
-    private Long categoryId;     // Foreign key to category
+    private BigInteger categoryId;     // Foreign key to category
     private Integer price;       // Product price
-    private Double discount;     // Discount percentage
+    private float discount;     // Discount percentage
 
     public String getSku() { return sku; }
     public void setSku(String sku) { this.sku = sku; }
-    public Long getCategoryId() { return categoryId; }
-    public void setCategoryId(Long categoryId) { this.categoryId = categoryId; }
+    public BigInteger getCategoryId() { return categoryId; }
+    public void setCategoryId(BigInteger categoryId) { this.categoryId = categoryId; }
     public Integer getPrice() { return price; }
     public void setPrice(Integer price) { this.price = price; }
-    public Double getDiscount() { return discount; }
-    public void setDiscount(Double discount) { this.discount = discount; }
+    public float getDiscount() { return discount; }
+    public void setDiscount(float discount) { this.discount = discount; }
+
 }
